Various fixes to items, ore gen
Added all the ore chunk items and the correspondig loot tables for the drops
Lang file has been updated with the name of the blocks and items added in past pulls
Implemented ore rarity
Separated the item and ore classes
Fixed texture and item name errors, fixed platnium named as "platinum"
Ore spawn balanced to spawn much less ores than before and another minor fixes
Removed ModItems and ModBlocks classes, models, items and lang definitions
Diamond, Lignite and Native Gold has been removed as well
MiscItems class created and added Multimeter item to it
Refined OreGen to include vein size and depth
Created MiscBlocks class and added the Flubber to it renamed as "Flubber_block"

+++ integration-mc1-16/src/main/kotlin/org/eln2/utils/OreData.kt
@@ -8,6 +8,12 @@ import net.minecraft.item.ItemStack
  *
  * @param hardness How hard is the block (pickaxe type basically)
  * @param name Name of the ore (eg, "native_copper")
+ * @param rarity Rarity of the ore (eg, ->
+ * Very Rare 1 (times 2 for ore vein size, times 25 for world spawn height)
+ * Rare 2
+ * Uncommon 3
+ * Common 4
+ * Very Common 5)
  * @param result The item to drop
  */
-data class OreData(val hardness: Float, val name: String, val rarity: Float, val result: ItemStack)
+data class OreData(val hardness: Float, val name: String, val rarity: Int, val result: ItemStack)
